split card states
