+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/settings/mealssettings/ui/MealSettingsCardStateWithMeal.kt
@@ -1,19 +1,11 @@
 package com.maksimowiczm.foodyou.feature.settings.mealssettings.ui
 
-import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.Stable
 import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.text.input.TextFieldValue
 import com.maksimowiczm.foodyou.data.model.Meal
-import kotlin.time.Duration.Companion.hours
-import kotlinx.datetime.Clock
-import kotlinx.datetime.TimeZone
-import kotlinx.datetime.toLocalDateTime
 
 interface MealSettingsCardState {
     val nameInput: MutableState<TextFieldValue>
@@ -60,54 +52,5 @@ class MealSettingsCardStateWithMeal(
             to = toTime,
             rank = meal.rank
         )
-    }
-}
-
-@Composable
-fun rememberMealSettingsCardState(): MealSettingsCardState {
-    val nameInput = rememberSaveable(
-        stateSaver = TextFieldValue.Saver
-    ) {
-        mutableStateOf(TextFieldValue())
-    }
-
-    val from = Clock.System.now().toLocalDateTime(TimeZone.currentSystemDefault())
-    val to = Clock.System.now().plus(2.hours).toLocalDateTime(TimeZone.currentSystemDefault())
-
-    val fromTimeInput = rememberSaveable(
-        saver = LocalTimeInput.Saver
-    ) {
-        LocalTimeInput(from.time)
-    }
-
-    val toTimeInput = rememberSaveable(
-        saver = LocalTimeInput.Saver
-    ) {
-        LocalTimeInput(to.time)
-    }
-
-    val isAllDay = rememberSaveable {
-        mutableStateOf(false)
-    }
-
-    return remember {
-        MealSettingsCardStateWithoutMeal(
-            nameInput = nameInput,
-            fromTimeInput = fromTimeInput,
-            toTimeInput = toTimeInput,
-            isAllDay = isAllDay
-        )
-    }
-}
-
-@Stable
-class MealSettingsCardStateWithoutMeal(
-    override val nameInput: MutableState<TextFieldValue>,
-    override val fromTimeInput: LocalTimeInput,
-    override val toTimeInput: LocalTimeInput,
-    override val isAllDay: MutableState<Boolean>
-) : MealSettingsCardState {
-    override val isDirty by derivedStateOf {
-        nameInput.value.text.isNotEmpty()
     }
 }
